GUI/Presenter: Fixed small bug in View class. ViewBaseTk's "View" menu can no longer be torn off.
The GuiBlock in GuiBlockTest.py seems to disappear when the window is resized specifically because the _resize_view method calls the redraw() method, which deletes every item on the canvas before redrawing just the background. Nothing is being invoked specifically to redraw the GuiBlock itself. Need to look over the Perl code once more to verify whether this is as intended; if so, then the responsibility of redrawing the GuiBlock after resizing the window lies elsewhere.