GUI/Presenter: Fixed bugs in Scheduler and AllScheduables.
Can now get past previous issue with Notebook tabs. Now tab creation is crashing because Pmw.ScrolledFrame doesn't have a tk attribute, meaning it doesn't seem to be compatible with certain Tk functions.

The crash occurs in ViewsManagerTk.create_buttons_for_frame, when it tries to create a button on the ScrolledFrame object.

The offending ScrolledFrame is created in SchedulerTk.draw_view_choices.