GUI/Presenter: Changed view_choices_scrolled_frame to a scrolled(Frame) object. Solves previous bug.
Results in a new problem. When ViewsManagerTk calls create_buttons_for_frame for the scrolled(Frame), it crashes because some of the items were created with the grid geometry manager, while the parent already had child items created with pack. So anything grid-based that shares its parent with these things needs to be changed to pack for consistency.