GUI: Fixed bugs in ViewTk callbacks re: getting coordinates of canvas items and # of arguments passed to callback functions.
Attempting to move the block currently fails because the x and y coordinates are not accurately reflecting the position of the mouse when _gui_block_has_stopped_moving() is called. It thinks that the mouse is somewhere off the left edge of the canvas/window, meaning it tries to assign a day which doesn't exist to the GuiBlock's underlying Block.