GUI: Beginning to fix a bug in the selection of AssignBlocks.
Partial fix: user no longer automatically selects four or five blocks when clicking on a blank space in the grid, and can manually select the number of half-hour blocks they want. However, the AssignToResource window shows that the block it would create has absurdly high durations: 90+ hours for a block that should only be two.